refactor: abstract Blackjack, Bust, Stay state to Finished
